requirements.txt
flask
python-docx
sentence-transformers
chromadb
requests
python-dotenv
